% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_cluster.R
\name{layout_cluster_panel}
\alias{layout_cluster_panel}
\title{layout_cluster_panel}
\usage{
layout_cluster_panel(
  g,
  cluster,
  nrow = 2,
  per_row = NULL,
  per_layout = "circle",
  x_space = 1,
  y_space = 1,
  heights = NULL
)
}
\arguments{
\item{g}{tbl_graph object}

\item{cluster}{cluster column in node table}
}
\value{
matrix of layout
}
\description{
layout_cluster_panel
}
