% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_cluster_col.R
\name{layout_cluster_panel_col}
\alias{layout_cluster_panel_col}
\title{layout_cluster_panel_col}
\usage{
layout_cluster_panel_col(
  g,
  cluster,
  ncol = 2,
  per_col = NULL,
  per_layout = "circle",
  x_space = 1,
  y_space = 1,
  widths = NULL
)
}
\arguments{
\item{g}{tbl_graph object}

\item{cluster}{cluster column in node table}
}
\value{
matrix of layout
}
\description{
layout_cluster_panel_col
}
