% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_cluster_wise.R
\name{layout_cluster_wise}
\alias{layout_cluster_wise}
\title{layout_cluster_wise}
\usage{
layout_cluster_wise(graph, cluster_col, per_layout = layout_in_circle)
}
\value{
matrix of layout
}
\description{
split to subgraph, use merge_coords to merge the layouts
based on dla algorithm
}
