% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_with_explicit_edge_weight.R
\name{layout_with_explicit_edge_weight}
\alias{layout_with_explicit_edge_weight}
\title{layout_with_explicit_edge_weight}
\usage{
layout_with_explicit_edge_weight(
  g,
  cluster,
  use_layout = layout_with_fr,
  within = 10,
  between = 1,
  original_weight = FALSE
)
}
\arguments{
\item{g}{tbl_graph object}

\item{cluster}{cluster column in node table}

\item{use_layout}{should have weights argument}

\item{original_weight}{reset original weights}
}
\value{
matrix of layout
}
\description{
This is based on the stackoverflow answers
https://stackoverflow.com/questions/16390221/how-to-make-grouped-layout-in-igraph
}
